+++ src/middleware.rs
@@ -1,9 +1,3 @@
-#[cfg(feature = "brotli")]
-use async_compression::futures::bufread::BrotliEncoder;
-#[cfg(feature = "deflate")]
-use async_compression::futures::bufread::DeflateEncoder;
-#[cfg(feature = "gzip")]
-use async_compression::futures::bufread::GzipEncoder;
 use futures_lite::io::BufReader;
 use tide::http::cache::{CacheControl, CacheDirective};
 use tide::http::conditional::Vary;
@@ -11,7 +5,27 @@ use tide::http::content::{AcceptEncoding, ContentEncoding, Encoding};
 use tide::http::{headers, Body, Method};
 use tide::{Middleware, Next, Request, Response};
 
+#[cfg(feature = "brotli")]
+use async_compression::futures::bufread::BrotliEncoder;
+#[cfg(feature = "deflate")]
+use async_compression::futures::bufread::DeflateEncoder;
+#[cfg(feature = "gzip")]
+use async_compression::futures::bufread::GzipEncoder;
+
+#[cfg(feature = "regex-check")]
+use http_types::content::ContentType;
+#[cfg(feature = "regex-check")]
+use regex::{Regex, RegexBuilder};
+
 const THRESHOLD: usize = 1024;
+
+// These regular expressions ere taken from jshttp/compressible
+// Used under terms of the MIT license.
+// https://github.com/jshttp/compressible/blob/89b61014fb82f0c64b42acef12d161dee48fb58e/index.js#L23-L24
+#[cfg(feature = "regex-check")]
+const CONTENT_TYPE_CHECK_PATTERN: &str = r"^text/|\+(?:json|text|xml)$";
+#[cfg(feature = "regex-check")]
+const EXTRACT_TYPE_PATTERN: &str = r"^\s*([^;\s]*)(?:;|\s|$)";
 
 /// A middleware for compressing response body data.
 ///
@@ -26,20 +40,24 @@ const THRESHOLD: usize = 1024;
 #[derive(Clone, Debug)]
 pub struct CompressMiddleware {
     threshold: usize,
+    #[cfg(feature = "regex-check")]
+    content_type_check: Option<Regex>,
+    #[cfg(feature = "regex-check")]
+    extract_type_regex: Regex,
 }
 
 impl Default for CompressMiddleware {
     fn default() -> Self {
-        CompressMiddleware {
-            threshold: THRESHOLD,
-        }
+        CompressMiddlewareBuilder::default().into()
     }
 }
 
 impl CompressMiddleware {
     /// Creates a new CompressMiddleware.
     ///
-    /// Uses the default minimum body size threshold (1024 bytes).
+    /// Uses the defaults:
+    /// - Minimum body size threshold (1024 bytes).
+    /// - Check for `Content-Type` header match `^text/|\+(?:json|text|xml)$` (case insensitive).
     ///
     /// ## Example
     /// ```rust
@@ -53,22 +71,33 @@ impl CompressMiddleware {
         Self::default()
     }
 
-    /// Creates a new CompressMiddleware with a custom minimum body size threshold.
+    /// Used to create a new CompressMiddleware with custom settings.
     ///
-    /// # Arguments
-    ///
-    /// * `threshold` - minimum body size in bytes.
-    ///
-    /// ## Example
-    /// ```rust
-    /// # async_std::task::block_on(async {
-    /// let mut app = tide::new();
-    ///
-    /// app.with(tide_compress::CompressMiddleware::with_threshold(512));
-    /// # })
-    /// ```
-    pub fn with_threshold(threshold: usize) -> Self {
-        CompressMiddleware { threshold }
+    /// See [`CompressMiddlewareBuilder`]
+    pub fn builder() -> CompressMiddlewareBuilder {
+        CompressMiddlewareBuilder::new()
+    }
+
+    /// Sets the minimum body size threshold value.
+    pub fn set_threshold(&mut self, threshold: usize) {
+        self.threshold = threshold
+    }
+
+    /// Gets the existing minimum body size threshold value.
+    pub fn threshold(&self) -> usize {
+        self.threshold
+    }
+
+    #[cfg(feature = "regex-check")]
+    /// Sets the `Content-Type` header (MIME) check regular expression.
+    pub fn set_content_type_check(&mut self, content_type_check: Option<Regex>) {
+        self.content_type_check = content_type_check
+    }
+
+    #[cfg(feature = "regex-check")]
+    /// Gets a reference to the existing `Content-Type` header (MIME) check regular expression.
+    pub fn content_type_check(&self) -> Option<&Regex> {
+        self.content_type_check.as_ref()
     }
 }
 
@@ -122,6 +151,22 @@ impl<State: Clone + Send + Sync + 'static> Middleware<State> for CompressMiddlew
             }
         }
 
+        #[cfg(feature = "regex-check")]
+        // Check if the `Content-Type` header indicates a compressible body.
+        if let Some(ref content_type_check) = self.content_type_check {
+            if let Some(content_type) = ContentType::from_headers(&res)? {
+                if let Some(extension_match) = self
+                    .extract_type_regex
+                    .captures(content_type.value().as_str())
+                    .and_then(|captures| captures.get(1))
+                {
+                    if !content_type_check.is_match(extension_match.as_str()) {
+                        return Ok(res);
+                    }
+                }
+            }
+        }
+
         let body = res.take_body();
         let encoding = accepts.negotiate(&[
             #[cfg(feature = "brotli")]
@@ -167,4 +212,86 @@ fn get_encoder(body: Body, encoding: &ContentEncoding) -> Body {
     }
 
     body
+}
+
+#[derive(Clone, Debug)]
+/// Used to create a new CompressMiddleware with custom settings.
+///
+/// Uses the defaults:
+/// - Minimum body size threshold (1024 bytes).
+/// - Check for `Content-Type` header match `^text/|\+(?:json|text|xml)$` (case insensitive).
+///
+/// ## Example
+/// ```rust
+/// # async_std::task::block_on(async {
+/// let mut app = tide::new();
+///
+/// let check_regex = regex::Regex::new(r"^text/|\+(?:json|text|xml)$").expect("regular expression defined in source code");
+///
+/// let compress_middleware = tide_compress::CompressMiddleware::builder()
+///     .threshold(1024)
+///     .content_type_check(Some(check_regex))
+///     .build();
+///
+/// app.with(compress_middleware);
+/// # })
+/// ```
+pub struct CompressMiddlewareBuilder {
+    pub threshold: usize,
+    #[cfg(feature = "regex-check")]
+    pub content_type_check: Option<Regex>,
+}
+
+impl Default for CompressMiddlewareBuilder {
+    fn default() -> Self {
+        Self {
+            threshold: THRESHOLD,
+            #[cfg(feature = "regex-check")]
+            content_type_check: Some(
+                RegexBuilder::new(CONTENT_TYPE_CHECK_PATTERN)
+                    .case_insensitive(true)
+                    .build()
+                    .expect("Constant regular expression defined in Tide-Compress's source code"),
+            ),
+        }
+    }
+}
+
+impl CompressMiddlewareBuilder {
+    /// Make a new builder.
+    /// Identical to `CompressMiddleware::builder()`.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Sets the minimum body size threshold value.
+    pub fn threshold(mut self, threshold: usize) -> Self {
+        self.threshold = threshold;
+        self
+    }
+
+    #[cfg(feature = "regex-check")]
+    /// Sets the `Content-Type` header (MIME) check regular expression.
+    pub fn content_type_check(mut self, content_type_check: Option<Regex>) -> Self {
+        self.content_type_check = content_type_check;
+        self
+    }
+
+    /// Construct a middleware instance from this builder.
+    pub fn build(self) -> CompressMiddleware {
+        self.into()
+    }
+}
+
+impl From<CompressMiddlewareBuilder> for CompressMiddleware {
+    fn from(builder: CompressMiddlewareBuilder) -> Self {
+        Self {
+            threshold: builder.threshold,
+            #[cfg(feature = "regex-check")]
+            content_type_check: builder.content_type_check,
+            #[cfg(feature = "regex-check")]
+            extract_type_regex: Regex::new(EXTRACT_TYPE_PATTERN)
+                .expect("Constant regular expression defined in Tide-Compress's source code"),
+        }
+    }
 }