+++ tests/compressed.rs
@@ -21,7 +21,11 @@ const BR_COMPRESSED: &[u8] = &[
 #[async_std::test]
 async fn brotli_compressed() {
     let mut app = tide::new();
-    app.with(tide_compress::CompressMiddleware::with_threshold(16));
+    app.with(
+        tide_compress::CompressMiddleware::builder()
+            .threshold(16)
+            .build(),
+    );
     app.at("/").get(|_| async {
         let mut res = Response::new(StatusCode::Ok);
         res.set_body(TEXT.to_owned());
@@ -69,7 +73,11 @@ const GZIPPED: &[u8] = &[
 #[async_std::test]
 async fn gzip_compressed() {
     let mut app = tide::new();
-    app.with(tide_compress::CompressMiddleware::with_threshold(16));
+    app.with(
+        tide_compress::CompressMiddleware::builder()
+            .threshold(16)
+            .build(),
+    );
     app.at("/").get(|_| async move {
         let mut res = Response::new(StatusCode::Ok);
         res.set_body(TEXT.to_owned());
@@ -99,7 +107,11 @@ const DEFLATED: &'static [u8] = &[
 #[async_std::test]
 async fn deflate_compressed() {
     let mut app = tide::new();
-    app.with(tide_compress::CompressMiddleware::with_threshold(16));
+    app.with(
+        tide_compress::CompressMiddleware::builder()
+            .threshold(16)
+            .build(),
+    );
     app.at("/").get(|_| async {
         let mut res = Response::new(StatusCode::Ok);
         res.set_body(TEXT.to_owned());