+++ Changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Feat: `Content-Type` header compressibility check via `Regex`.
+
 ## [0.9.0] - 2021-01-29
 
 - Enabled `DEFLATE` support by default, since there is no real drawbacks to doing so.